Upgrade deps available to reduce vulnerabilities
