+++ .github/workflows/build.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ${{ matrix.os }}
 
     env:
-      node-version: 10
+      node-version: 14
 
     strategy:
       matrix:
