+++ package.json
@@ -112,10 +112,10 @@
     "@babel/preset-react": "^7.13.13",
     "@babel/preset-typescript": "^7.13.0",
     "@storybook/addon-actions": "^6.3.0-alpha.33",
-    "@storybook/addon-essentials": "^6.3.0-alpha.33",
+    "@storybook/addon-essentials": "^6.3.0-alpha.41",
     "@storybook/addon-links": "^6.3.0-alpha.33",
     "@storybook/builder-webpack5": "^6.3.0-alpha.33",
-    "@storybook/react": "^6.3.0-alpha.33",
+    "@storybook/react": "^6.3.0-alpha.41",
     "@types/chai": "^4.2.15",
     "@types/electron": "^1.6.10",
     "@types/enzyme": "^3.10.8",
@@ -131,7 +131,7 @@
     "@types/webpack-env": "^1.16.0",
     "@typescript-eslint/eslint-plugin": "^4.16.1",
     "@typescript-eslint/parser": "^4.16.1",
-    "autoprefixer": "^9.8.0",
+    "autoprefixer": "^10.2.5",
     "babel-loader": "^8.2.2",
     "brace": "^0.11.0",
     "breakpoint-sass": "^2.6.1",
@@ -140,12 +140,12 @@
     "concurrently": "^2.2.0",
     "cross-env": "^7.0.2",
     "csp-html-webpack-plugin": "^5.1.0",
-    "css-loader": "^4.2.1",
+    "css-loader": "^5.2.5",
     "csv-stringify": "^5.5.1",
     "dom-helpers": "^5.1.4",
     "dotenv-webpack": "^7.0.2",
     "electron": "^12.0.1",
-    "electron-builder": "^22.8.0",
+    "electron-builder": "^22.11.5",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "eslint": "^7.21.0",
@@ -161,7 +161,7 @@
     "html2canvas": "^0.5.0-beta4",
     "husky": "^5.1.1",
     "imports-loader": "^0.6.5",
-    "jointjs": "^2.2.1",
+    "jointjs": "^3.3.1",
     "jquery": "^3.3.1",
     "jsdom": "^16.5.1",
     "json-loader": "^0.5.4",
@@ -180,7 +180,8 @@
     "node-sass": "^5.0.0",
     "playwright-core": "^1.9.1",
     "playwright-electron": "^0.5.0",
-    "postcss-loader": "^3.0.0",
+    "postcss": "^8.3.0",
+    "postcss-loader": "^5.3.0",
     "prettier": "^2.2.1",
     "prop-types": "^15.7.2",
     "react": "^16.14.0",
